+++ Sources/MacIMEKit/Args.swift
@@ -147,6 +147,12 @@ public enum ArgsIMED {
          case "--help", "-h":
             IO.out(Help.macimed)
             exit(0)
+         case "--info", "-i":
+            var info: [String] = []
+            info.append("sock: \(state.sockPath ?? "")")
+            info.append("macime: \(state.macimePath ?? "")")  // Does not reflect running macimed's macimePath when executed by `MACIME_PATH="xxx"; macimed`
+            IO.out(info.joined(separator: "\n"))
+            exit(0)
          default:
             break
          }